+++ salon-frontend/src/pages/Calendar/calendarSchedule.tsx
@@ -3,7 +3,7 @@ import FullCalendar from "@fullcalendar/react";
 import dayGridPlugin from "@fullcalendar/daygrid";
 import timeGridPlugin from "@fullcalendar/timegrid";
 import interactionPlugin from "@fullcalendar/interaction";
-import { INITIAL_EVENTS, createEventId } from "../lib/event";
+import { INITIAL_EVENTS, createEventId } from "../../lib/event";
 import { useGetAllAppointments } from "@/lib/hooks/getAllAppointments";
 import { useUser } from "@clerk/clerk-react";
 import { useNavigate } from "react-router-dom";
