+++ salon-frontend/src/pages/Landing/signinPage.tsx
@@ -1,4 +1,4 @@
-import UnlockedNavbar from "@/modules/unlockedNavbar";
+import UnlockedNavbar from "@/pages/Navbar/unlockedNavbar";
 import { SignIn } from "@clerk/clerk-react";
 
 export default function SignInPage() {
