+++ salon-frontend/src/pages/Messages/Chat/chatTerminal.tsx
