+++ salon-frontend/src/pages/Navbar/navbar.tsx
