+++ salon-frontend/src/pages/Navbar/unlockedNavbar.tsx
